+++ packages/jaspr_lints/CHANGELOG.md
@@ -5,6 +5,7 @@
 ## 0.6.1
 
 - Allow `analyzer` versions 10.x.
+- Enabled `addStyles` and `convertToNestedStyles` assists in client mode projects. 
 
 ## 0.6.0